requirements.txt
amqp==2.5.2
arrow==0.15.6
asgiref==3.2.7
asn1crypto==0.24.0
billiard==3.6.3.0
blessed==1.17.6
blessings==1.7
braintree==3.52.0
celery==4.4.2
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.6.1
defusedxml==0.5.0
dj-paypal==0.12.0
dj-stripe==2.0.0
Django==3.0.7
django-allauth==0.39.1
django-appconf==1.0.3
django-bootstrap-modal-forms==1.4.2
django-common-helpers==0.9.2
django-crispy-forms==1.7.2
django-cron==0.5.1
django-debug-toolbar==2.0
django-ip-shield==0.1.7
django-jquery==3.1.0
django-js-asset==1.2.2
django-mptt==0.9.1
django-password-reset==2.0
django-payments==0.13.0
django-picklefield==2.1.1
django-q==1.2.3
django-recaptcha3==0.1.0
django-redis==4.12.1
django-select2==7.1.0
django-select2-forms==2.1.0
django-sortedm2m==1.5.0
django-static-jquery==2.1.4
django-stronghold==0.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
fuzzywuzzy==0.17.0
guppy3==3.0.9
idna==2.8
importlib-metadata==1.6.0
inquirer==2.5.1
jsonfield==2.0.2
kombu==4.6.8
mysqlclient==1.4.2.post1
numpy==1.17.0
oauthlib==3.0.1
opencv-contrib-python==4.1.0.25
paypalrestsdk==1.13.1
Pillow==6.2.2
pipenv==2018.11.26
psycopg2==2.8.5
pycparser==2.19
Pygments==2.5.2
pyguetzli==1.0.8
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-Levenshtein==0.12.0
python3-openid==3.1.0
pytz==2018.9
readchar==2.0.1
redis==3.5.3
regex==2020.2.20
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
stripe==2.24.1
stronghold==1.5
suds-jurko==0.6
urllib3==1.24.2
vine==1.3.0
virtualenv==16.4.3
virtualenv-clone==0.5.3
wcwidth==0.2.3
xmltodict==0.12.0
zipp==3.1.0
